requirements.txt
transformers
torch
sentencepiece
colorama
python-telegram-bot==13.15
sacremoses
